Fixing lint errors

+++ tests/electron/spec.js
@@ -1,4 +1,20 @@
-"use strict";
+////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2021 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////
 
 const assert = require("assert");
 const path = require("path");
@@ -8,7 +24,7 @@ const Realm = require("realm");
 const RealmTests = require("realm-tests");
 
 describe("Test harness", () => {
-  if(global.options && global.options.runIn === "main") {
+  if (global.options && global.options.runIn === "main") {
     it("runs the test in the main process", () => {
       assert(process.versions.chrome, "Expected a chrome version");
       assert(!global.window, "Expected no window constant");
@@ -33,7 +49,7 @@ describe("Test harness", () => {
 
   it("loads Realm", () => {
     assert(Realm);
-    assert.equal(typeof(Realm), "function");
+    assert.equal(typeof Realm, "function");
     assert.equal(Realm.name, "Realm");
   });
 
@@ -49,8 +65,8 @@ describe("Test harness", () => {
 // Setting the timeout to the same as the ../../spec/unit_tests.js
 jasmine.DEFAULT_TIMEOUT_INTERVAL = 30000;
 
-Realm.copyBundledRealmFiles = function() {
-  const sourceDir = path.join(__dirname, '../data');
+Realm.copyBundledRealmFiles = function () {
+  const sourceDir = path.join(__dirname, "../data");
   const destinationDir = path.dirname(Realm.defaultPath);
 
   for (let filename of fs.readdirSync(sourceDir)) {
@@ -58,22 +74,19 @@ Realm.copyBundledRealmFiles = function() {
     let dest = path.join(destinationDir, filename);
 
     // If the destination file already exists, then don't overwrite it.
-    try {
-        fs.accessSync(dest);
-        continue;
-    } catch (e) {}
-
-    fs.writeFileSync(dest, fs.readFileSync(src));
+    const exists = fs.existsSync(dest);
+    if (!exists) {
+      fs.writeFileSync(dest, fs.readFileSync(src));
+    }
   }
 };
 
 const tests = RealmTests.getTestNames();
 for (const suiteName in tests) {
   describe(suiteName, () => {
+    beforeAll((done) => RealmTests.prepare(done));
 
-    beforeAll(done => RealmTests.prepare(done));
-
-    beforeEach(() => RealmTests.runTest(suiteName, 'beforeEach'));
+    beforeEach(() => RealmTests.runTest(suiteName, "beforeEach"));
 
     for (const testName of tests[suiteName]) {
       it(testName, (done) => {
@@ -90,6 +103,6 @@ for (const suiteName in tests) {
       });
     }
 
-    afterEach(() => RealmTests.runTest(suiteName, 'afterEach'));
+    afterEach(() => RealmTests.runTest(suiteName, "afterEach"));
   });
 }
